Update home design

+++ app/src/main/java/com/example/foodkit/model/FavoriteFood.kt
@@ -12,7 +12,7 @@ data class FavoriteFood(
     val imageUrl: String,
     val price: Double,
     val category: String,
-    val rating: Double,
+    val rating: Float,
     val isFavorite: Boolean,
     val numberRating: Int
 )
