+++ app/src/main/java/com/example/foodkit/presentation/viewModel/FavoriteFoodViewModel.kt
@@ -19,9 +19,27 @@ class FavoriteFoodViewModel(private val favoriteFoodDao: FavoriteFoodDao) : View
         }
     }
 
-    fun addFavoriteFood(name: String, description: String, userId: Int, imageUrl: String, price: Double, category: String, rating: Double) {
+    fun addFavoriteFood(
+        name: String,
+        description: String,
+        userId: Int,
+        imageUrl: String,
+        price: Double,
+        category: String,
+        rating: Float,
+    ) {
         viewModelScope.launch {
-            val newFood = FavoriteFood(name = name, description = description, userId = userId, imageUrl = imageUrl, price = price, category = category, rating = rating, isFavorite = true, numberRating = 0)
+            val newFood = FavoriteFood(
+                name = name,
+                description = description,
+                userId = userId,
+                imageUrl = imageUrl,
+                price = price,
+                category = category,
+                rating = rating,
+                isFavorite = true,
+                numberRating = 0
+            )
             favoriteFoodDao.insert(newFood)
             _favoriteFoods.value = favoriteFoodDao.getFavoriteFoodsForUser(userId)
         }
